fix, feat(orcamento, banco de dados): corrige tipo do atributo valorTotal para BigDecimal, adiciona migrations para criação das tabelas status_orcamentos e orcamentos - também popula a tabela status_orcamentos com os status da regra de negócio

+++ src/main/java/com/api/controleverbasbackend/dto/orcamento/DadosDetalhamentoOrcamento.java
@@ -1,5 +1,6 @@
 package com.api.controleverbasbackend.dto.orcamento;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.api.controleverbasbackend.domain.orcamento.Orcamento;
@@ -9,7 +10,7 @@ public record DadosDetalhamentoOrcamento(
         String fornecedor,
         String descricao,
         String formaPagamento,
-        Double valorTotal,
+        BigDecimal valorTotal,
         String observacoesGerais,
         String solicitanteNome,
         String gestorNome,
